refactor(dashboard-rework): Moved the reward query down
