+++ apps/website/src/app/accounts/[tag]/dashboard/(overview)/page.tsx
@@ -7,14 +7,13 @@ import { getChestCountPerEvent, getPossibleChestCount } from "@/actions/event";
 import { getAccountByTag } from "@/actions/account";
 import LatestChestCard from "@/components/latest-chest-card";
 import TotalChestCard from "@/components/total-chest-card";
-import { getChestCountPerReward } from "@/actions/reward";
 import ChestCountEventTable from "@/components/chest-count-event-table";
-import ChestCountRewardChart from "@/components/chest-count-reward-chart";
 import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
 import AccountInformationItem from "@/components/account-information-item";
 import { FilterConfig } from "@/types/common";
 import RarityCard from "@/components/rarity-card";
 import CategoryCard from "@/components/category-card";
+import RewardCard from "@/components/reward-card";
 
 export default async function Page({
   params,
@@ -33,7 +32,6 @@ export default async function Page({
   const possibleChestCount = await getPossibleChestCount();
   const chestCount = await getTotalChests({ accountId: account.id });
   const events = await getChestCountPerEvent({ accountId: account.id });
-  const rewards = await getChestCountPerReward({ accountId: account.id });
   const peakOpeningHourData = await getPeakOpeningHourData({
     accountId: account.id,
   });
@@ -56,7 +54,7 @@ export default async function Page({
         <RarityCard filters={filters} />
         <CategoryCard filters={filters} />
       </div>
-      <ChestCountRewardChart rewards={rewards} />
+      <RewardCard filters={filters} />
       <ChestCountEventTable events={events} accountCount={1} />
     </div>
   );
