+++ apps/website/src/app/events/[code]/dashboard/(overview)/page.tsx
@@ -5,9 +5,7 @@ import {
   getTotalChests,
 } from "@/actions/chest";
 import { getEventByCode } from "@/actions/event";
-import { getChestCountPerReward } from "@/actions/reward";
 import ChestCountAccountChart from "@/components/chest-count-account-chart";
-import ChestCountRewardChart from "@/components/chest-count-reward-chart";
 import LatestChestCard from "@/components/latest-chest-card";
 import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
 import TotalChestCard from "@/components/total-chest-card";
@@ -16,6 +14,7 @@ import EventNotFound from "@/components/event-not-found";
 import RarityCard from "@/components/rarity-card";
 import { FilterConfig } from "@/types/common";
 import CategoryCard from "@/components/category-card";
+import RewardCard from "@/components/reward-card";
 
 export default async function Page({
   params,
@@ -44,9 +43,6 @@ export default async function Page({
   const peakOpeningHourData = await getPeakOpeningHourData({
     eventId: event.id,
   });
-  const rewards = await getChestCountPerReward({
-    eventId: event.id,
-  });
   const accounts = await getChestCountPerAccount({
     eventId: event.id,
   });
@@ -69,7 +65,7 @@ export default async function Page({
         <RarityCard filters={filters} />
         <CategoryCard filters={filters} />
       </div>
-      <ChestCountRewardChart rewards={rewards} />
+      <RewardCard filters={filters} />
       <ChestCountAccountChart accounts={accounts} />
     </div>
   );
